+++ .eslintrc.json
@@ -164,7 +164,6 @@
     "line-comment-position": ["error", { "position": "above" }],
     // recommended rules overwrite
     "indent": "off",
-    "linebreak-style": ["error", "unix"],
     "newline-before-return": "error",
     "no-undef": "error",
     "padded-blocks": ["error", "never"],
