+++ app/admin/users/layout.tsx
@@ -1,5 +1,9 @@
 import React from "react"
 import {AdminLayout} from "@/components/layout/admin-layout"
+import {getServerSession} from "@/auth"
+import {redirect} from "next/navigation"
+import {createRedirectQueryParam} from "@/utils/url"
+import {profileService} from "@/services"
 
 export const dynamic = "force-dynamic"
 
@@ -8,6 +12,22 @@ export default async function LayoutAdminUserPage({
                                               }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const session = await getServerSession()
+
+  if (!session) {
+    redirect(`/login${createRedirectQueryParam(`/admin/users`)}`)
+  }
+
+  const userAndProfile = await profileService.findByUserUid(session.uid)
+
+  if (!userAndProfile) {
+    redirect(`/login${createRedirectQueryParam(`/admin/users`)}`)
+  }
+
+  if (userAndProfile.role !== "admin") {
+    redirect(`/`)
+  }
+
   return (
     <>
       <AdminLayout>
